+++ src/main/java/com/ivan/labdb4/model/Highlight.java
@@ -1,6 +1,6 @@
 package com.ivan.labdb4.model;
 
-import jakarta.persistence.*;
+import javax.persistence.*;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -11,7 +11,7 @@ public class Highlight {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     @OneToOne(targetEntity = HighlightAuthor.class, fetch = FetchType.EAGER, optional = false)
     @JoinColumn(name = "author_id", referencedColumnName = "id")
     private HighlightAuthor author;
