+++ src/main/java/com/ivan/labdb4/repository/HighlightAuthorRepository.java
@@ -3,6 +3,6 @@ package com.ivan.labdb4.repository;
 import com.ivan.labdb4.model.HighlightAuthor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HighlightAuthorRepository extends JpaRepository<HighlightAuthor, Long> {
+public interface HighlightAuthorRepository extends JpaRepository<HighlightAuthor, Integer> {
 
 }
