+++ src/main/java/com/ivan/labdb4/repository/HighlightRepository.java
@@ -3,5 +3,5 @@ package com.ivan.labdb4.repository;
 import com.ivan.labdb4.model.Highlight;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HighlightRepository extends JpaRepository<Highlight, Long> {
+public interface HighlightRepository extends JpaRepository<Highlight, Integer> {
 }
