+++ src/main/java/com/ivan/labdb4/repository/MovieMetaInfoRepository.java
@@ -3,5 +3,5 @@ package com.ivan.labdb4.repository;
 import com.ivan.labdb4.model.MovieMetaInfo;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MovieMetaInfoRepository extends CrudRepository<MovieMetaInfo, Long> {
+public interface MovieMetaInfoRepository extends CrudRepository<MovieMetaInfo, Integer> {
 }
